refactoring into services.ts
